[backend] identify non-idle workers for constraints and store in :packstatus.finished

+++ src/backend/BSSched/Checker.pm
@@ -848,7 +848,26 @@ sub checkpkgs {
     'packstatus' => \%packstatus,
     'packerror' => \%packerror,
   });
-  unlink("$gdst/:packstatus.finished");
+  # prune packstatus.finished
+  if (%building) {
+    my $psf = readstr("$gdst/:packstatus.finished", 1);
+    if ($psf) {
+      my %dispatchdetails;
+      for (split("\n", $psf)) {
+        my ($code, $rest) = split(' ', $_, 2);
+        next unless $code eq 'scheduled';
+        my ($packid, $job, $details) = split('/', $rest, 3);
+        $dispatchdetails{$packid} = "$_\n" if $job && ($building{$packid} || '') eq $job;
+      }
+      if (%dispatchdetails) {
+        writestr("$gdst/.:packstatus.finished", "$gdst/.:packstatus.finished", join('', sort values %dispatchdetails));
+      } else {
+        unlink("$gdst/:packstatus.finished");
+      }
+    }
+  } else {
+    unlink("$gdst/:packstatus.finished");
+  }
   my $schedulerstate;
   if (keys %building) {
     $schedulerstate = 'building';
